+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+* `onRefresh` and `onLoading` can receive `EasyRefreshType` and execute calls
+* `onRefresh` and `onLoading` adds `EasyRefreshController` callback parameter
+* `onLoading` change to `onLoad`
+
 ## 0.0.3
 
 * Fix the error of `placeholderFill`=true in specific situations